Minor restructuring of module
- moved RLAgent and ILAgent definition classes up
- moved connector package up
