+++ src/rl_framework/agent/__init__.py
@@ -1,8 +1,5 @@
 from .base_agent import Agent
-from .imitation import ILAgent, ImitationAgent
-from .reinforcement import (
-    AsyncStableBaselinesAgent,
-    CustomAgent,
-    RLAgent,
-    StableBaselinesAgent,
-)
+from .imitation import ImitationAgent
+from .imitation_learning_agent import ILAgent
+from .reinforcement import AsyncStableBaselinesAgent, CustomAgent, StableBaselinesAgent
+from .reinforcement_learning_agent import RLAgent
