+++ src/rl_framework/agent/reinforcement/__init__.py
@@ -1,4 +1,3 @@
 from .async_stable_baselines import AsyncStableBaselinesAgent
 from .custom_agent import CustomAgent
-from .reinforcement_learning_agent import RLAgent
 from .stable_baselines import StableBaselinesAgent
