+++ src/rl_framework/util/connector/hugging_face_connector.py
@@ -171,7 +171,8 @@ See further agent and evaluation metadata in the according README section.
 
 
 ## Import
-The Python module used for training and uploading/downloading is [reinforcement-learning-framework](https://github.com/alexander-zap/reinforcement-learning-framework).
+The Python module used for training and uploading/downloading is
+[reinforcement-learning-framework](https://github.com/alexander-zap/reinforcement-learning-framework).
 It is an easy-to-read, plug-and-use Reinforcement Learning framework and provides standardized interfaces
 and implementations to various Reinforcement Learning methods and environments.
 
@@ -198,7 +199,8 @@ agent.download(repository_id=repository_id, filename=file_name)
 
 ```
 
-Further examples can be found in the [exploration section of the reinforcement-learning-framework repository](https://github.com/alexander-zap/reinforcement-learning-framework/tree/main/exploration).
+Further examples can be found in the exploration section of the
+[reinforcement-learning-framework repository](https://github.com/alexander-zap/reinforcement-learning-framework).
 
             """
